requirements.txt
selenium
requests
PyYAML
webdriver-manager==4.0.0
python-decouple